add QuantestPyCircuit, rename TestCircuit->StateVectorCircuit, add converters from QPC to SVC or PC.
