+++ quantestpy/__init__.py
@@ -1,3 +1,4 @@
+from .simulator.quantestpy_circuit import QuantestPyCircuit
+from .simulator.state_vector_circuit import StateVectorCircuit
 from .simulator.pauli_circuit import PauliCircuit
-from .test_circuit import TestCircuit
 from .assertion.get_ctrl_val import assert_get_ctrl_val
