+++ quantestpy/simulator/exceptions.py
@@ -1,0 +1,10 @@
+class QuantestPyCircuitError(Exception):
+    pass
+
+
+class StateVectorCircuitError(Exception):
+    pass
+
+
+class PauliCircuitError(Exception):
+    pass
