+++ test/converter/qasm_and_qiskit/test_converter_without_qiskit.py
@@ -1,7 +1,8 @@
 import traceback
 import unittest
 
-from quantestpy.converter import _cvt_openqasm_to_test_circuit
+from quantestpy.converter.qasm_and_qiskit import \
+    _cvt_openqasm_to_quantestpy_circuit
 from quantestpy.exceptions import QuantestPyError
 
 
@@ -10,7 +11,8 @@ class TestConverter(unittest.TestCase):
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_converter_without_qiskit
+    $ python -m \
+        unittest test.converter.qasm_and_qiskit.test_converter_without_qiskit
     ...
     ----------------------------------------------------------------------
     Ran 5 tests in 0.041s
@@ -22,10 +24,10 @@ class TestConverter(unittest.TestCase):
         qasm = 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[2];\nx q[0];\n'
 
         with self.assertRaises(QuantestPyError):
-            _ = _cvt_openqasm_to_test_circuit(qasm)
+            _ = _cvt_openqasm_to_quantestpy_circuit(qasm)
 
         try:
-            _ = _cvt_openqasm_to_test_circuit(qasm)
+            _ = _cvt_openqasm_to_quantestpy_circuit(qasm)
 
         except QuantestPyError as e:
             expected_error_msg = "quantestpy.exceptions.QuantestPyError: " \
