+++ test/simulator/quantestpy_circuit/test_add_gate.py
@@ -1,26 +1,26 @@
 import traceback
 import unittest
 
-from quantestpy import TestCircuit
-from quantestpy.exceptions import QuantestPyTestCircuitError
+from quantestpy import QuantestPyCircuit
+from quantestpy.simulator.exceptions import QuantestPyCircuitError
 
 
-class TestTestCircuitAddGate(unittest.TestCase):
+class TestQuantestPyCircuitAddGate(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_test_circuit_add_gate
-    ................
+    $ python -m unittest test.simulator.quantestpy_circuit.test_add_gate
+    ............
     ----------------------------------------------------------------------
-    Ran 16 tests in 0.001s
+    Ran 12 tests in 0.001s
 
     OK
     $
     """
 
     def test_all_pass(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
         test_circuit.add_gate(
             {"name": "h",
                 "target_qubit": [1],
@@ -54,7 +54,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
         self.assertEqual(expected_gates, actual_gates)
 
     def test_all_pass_multi_target(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
         test_circuit.add_gate(
             {"name": "x",
                 "target_qubit": [1, 2],
@@ -88,7 +88,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
         self.assertEqual(expected_gates, actual_gates)
 
     def test_index_range_in_target_qubit(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
 
         try:
             self.assertIsNotNone(
@@ -101,9 +101,10 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: Index 3" \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
+                + "Index 3" \
                 + " in target_qubit out of range for test_circuit size 3.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -111,7 +112,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_index_type_in_target_qubit(self,):
-        test_circuit = TestCircuit(4)
+        test_circuit = QuantestPyCircuit(4)
 
         try:
             self.assertIsNotNone(
@@ -124,9 +125,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "Index in target_qubit must be integer type.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -134,7 +135,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_index_range_in_control_qubit(self,):
-        test_circuit = TestCircuit(10)
+        test_circuit = QuantestPyCircuit(10)
 
         try:
             self.assertIsNotNone(
@@ -147,9 +148,10 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: Index 10" \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
+                + "Index 10" \
                 + " in control_qubit out of range for test_circuit size 10.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -157,7 +159,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_index_type_in_control_qubit(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
 
         try:
             self.assertIsNotNone(
@@ -170,9 +172,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "Index in control_qubit must be integer type.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -180,7 +182,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_value_range_in_control_value(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
 
         try:
             self.assertIsNotNone(
@@ -193,9 +195,10 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: Value 2" \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
+                + "Value 2" \
                 + " in control_value is not acceptable. " \
                 + "It must be either 0 or 1.\n"
 
@@ -204,7 +207,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_value_type_in_control_value(self,):
-        test_circuit = TestCircuit(3)
+        test_circuit = QuantestPyCircuit(3)
 
         try:
             self.assertIsNotNone(
@@ -217,9 +220,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "Value in control_value must be integer type.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -227,7 +230,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
             self.assertEqual(expected_error_msg, actual_error_msg)
 
     def test_target_qubit_empty(self,):
-        test_circuit = TestCircuit(5)
+        test_circuit = QuantestPyCircuit(5)
 
         try:
             self.assertIsNotNone(
@@ -240,105 +243,10 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "'target_qubit' must not an empty list.\n"
-
-            actual_error_msg = traceback.format_exception_only(type(e), e)[0]
-
-            self.assertEqual(expected_error_msg, actual_error_msg)
-
-    def test_parameter_without_param(self,):
-        test_circuit = TestCircuit(1)
-
-        try:
-            self.assertIsNotNone(
-                test_circuit.add_gate(
-                    {"name": "x",
-                     "target_qubit": [0],
-                     "control_qubit": [],
-                     "control_value": [],
-                     "parameter": [1]}
-                )
-            )
-
-        except QuantestPyTestCircuitError as e:
-            expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
-                + "x gate must have " \
-                + "an empty list for 'parameter'.\n"
-
-            actual_error_msg = traceback.format_exception_only(type(e), e)[0]
-
-            self.assertEqual(expected_error_msg, actual_error_msg)
-
-    def test_parameter_with_one_param(self,):
-        test_circuit = TestCircuit(1)
-
-        try:
-            self.assertIsNotNone(
-                test_circuit.add_gate(
-                    {"name": "p",
-                     "target_qubit": [0],
-                     "control_qubit": [],
-                     "control_value": [],
-                     "parameter": []}
-                )
-            )
-
-        except QuantestPyTestCircuitError as e:
-            expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
-                + "p gate must have a list containing " \
-                + "exactly 1 element for 'parameter'.\n"
-
-            actual_error_msg = traceback.format_exception_only(type(e), e)[0]
-
-            self.assertEqual(expected_error_msg, actual_error_msg)
-
-    def test_parameter_with_four_param(self,):
-        test_circuit = TestCircuit(1)
-
-        try:
-            self.assertIsNotNone(
-                test_circuit.add_gate(
-                    {"name": "u",
-                     "target_qubit": [0],
-                     "control_qubit": [],
-                     "control_value": [],
-                     "parameter": []}
-                )
-            )
-
-        except QuantestPyTestCircuitError as e:
-            expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
-                + "u gate must have a list containing " \
-                + "exactly 4 elements for 'parameter'.\n"
-
-            actual_error_msg = traceback.format_exception_only(type(e), e)[0]
-
-            self.assertEqual(expected_error_msg, actual_error_msg)
-
-    def test_value_type_in_parameter(self,):
-        test_circuit = TestCircuit(1)
-
-        try:
-            self.assertIsNotNone(
-                test_circuit.add_gate(
-                    {"name": "p",
-                     "target_qubit": [0],
-                     "control_qubit": [],
-                     "control_value": [],
-                     "parameter": ["a"]}
-                )
-            )
-
-        except QuantestPyTestCircuitError as e:
-            expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
-                + "Parameter(s) in p gate must be float or integer type.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
 
@@ -346,7 +254,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
 
     def test_msg_from_duplicate_in_target_qubit(self,):
 
-        test_circuit = TestCircuit(5)
+        test_circuit = QuantestPyCircuit(5)
 
         try:
             self.assertIsNotNone(
@@ -359,9 +267,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "Duplicate in target_qubit is not supported.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -370,7 +278,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
 
     def test_msg_from_duplicate_in_control_qubit(self,):
 
-        test_circuit = TestCircuit(5)
+        test_circuit = QuantestPyCircuit(5)
 
         try:
             self.assertIsNotNone(
@@ -383,9 +291,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "Duplicate in control_qubit is not supported.\n"
 
             actual_error_msg = traceback.format_exception_only(type(e), e)[0]
@@ -394,7 +302,7 @@ class TestTestCircuitAddGate(unittest.TestCase):
 
     def test_msg_from_intersection_in_target_and_control_qubits(self,):
 
-        test_circuit = TestCircuit(5)
+        test_circuit = QuantestPyCircuit(5)
 
         try:
             self.assertIsNotNone(
@@ -407,9 +315,9 @@ class TestTestCircuitAddGate(unittest.TestCase):
                 )
             )
 
-        except QuantestPyTestCircuitError as e:
+        except QuantestPyCircuitError as e:
             expected_error_msg = \
-                "quantestpy.exceptions.QuantestPyTestCircuitError: " \
+                "quantestpy.simulator.exceptions.QuantestPyCircuitError: " \
                 + "target_qubit [0, 1, 2, 3] and control_qubit [3, 4] have " \
                 + "intersection.\n"
