+++ test/simulator/state_vector_circuit/test_cz_gate.py
@@ -2,109 +2,106 @@ import unittest
 
 import numpy as np
 
-from quantestpy import TestCircuit
-from quantestpy.test_circuit import _Y
+from quantestpy import StateVectorCircuit
+from quantestpy.simulator.state_vector_circuit import _Z
 
 
-class TestTestCircuitCYGate(unittest.TestCase):
+class TestStateVectorCircuitCZGate(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_test_circuit_cy_gate
+    $ python -m unittest test.simulator.state_vector_circuit.test_cz_gate
     ........
     ----------------------------------------------------------------------
-    Ran 7 tests in 0.008s
+    Ran 7 tests in 0.007s
 
     OK
     $
     """
 
-    def test_cy_regular_qubit_order(self,):
-        circ = TestCircuit(2)
+    def test_cz_regular_qubit_order(self,):
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[1], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[1], control_value=[1]
         )
 
         expected_gate = np.array([[1, 0, 0, 0],
                                   [0, 1, 0, 0],
-                                  [0, 0, 0, -1j],
-                                  [0, 0, 1j, 0]])
+                                  [0, 0, 1, 0],
+                                  [0, 0, 0, -1]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_qiskit_qubit_order(self,):
-        circ = TestCircuit(2)
+    def test_cz_qiskit_qubit_order(self,):
+        circ = StateVectorCircuit(2)
         circ._from_right_to_left_for_qubit_ids = True
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[1], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[1], control_value=[1]
         )
-        # this is qiskit's output
-        expected_gate = np.array([
-            [1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
-            [0.+0.j, 0.+0.j, 0.+0.j, 0.-1.j],
-            [0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j],
-            [0.+0.j, 0.+1.j, 0.+0.j, 0.+0.j]])
+
+        expected_gate = np.array([[1, 0, 0, 0],
+                                  [0, 1, 0, 0],
+                                  [0, 0, 1, 0],
+                                  [0, 0, 0, -1]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_flip_control_target(self,):
-        circ = TestCircuit(2)
+    def test_cz_flip_control_target(self,):
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[1], target_qubit=[0], control_value=[1]
+            _Z, control_qubit=[1], target_qubit=[0], control_value=[1]
         )
 
-        # this is qiskit's output
-        expected_gate = np.array([
-            [1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
-            [0.+0.j, 0.+0.j, 0.+0.j, 0.-1.j],
-            [0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j],
-            [0.+0.j, 0.+1.j, 0.+0.j, 0.+0.j]])
+        expected_gate = np.array([[1, 0, 0, 0],
+                                  [0, 1, 0, 0],
+                                  [0, 0, 1, 0],
+                                  [0, 0, 0, -1]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_three_qubits_qiskit_qubit_order(self,):
-        circ = TestCircuit(3)
+    def test_cz_three_qubits_qiskit_qubit_order(self,):
+        circ = StateVectorCircuit(3)
         circ._from_right_to_left_for_qubit_ids = True
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[2], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[2], control_value=[1]
         )
 
         # this is qiskit's output
         expected_gate = np.array([
             [1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
-            [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.-1.j, 0.+0.j, 0.+0.j],
+            [0.+0.j, 1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
             [0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
-            [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.-1.j],
+            [0.+0.j, 0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
             [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
-            [0.+0.j, 0.+1.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j],
+            [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, -1.+0.j, 0.+0.j, 0.+0.j],
             [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 1.+0.j, 0.+0.j],
-            [0.+0.j, 0.+0.j, 0.+0.j, 0.+1.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j]])
+            [0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, 0.+0.j, -1.+0.j]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_control_value_is_zero(self,):
-        circ = TestCircuit(2)
+    def test_cz_control_value_is_zero(self,):
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[1], control_value=[0]
+            _Z, control_qubit=[0], target_qubit=[1], control_value=[0]
         )
 
-        expected_gate = np.array([[0, -1j, 0, 0],
-                                  [1j, 0, 0, 0],
+        expected_gate = np.array([[1, 0, 0, 0],
+                                  [0, -1, 0, 0],
                                   [0, 0, 1, 0],
                                   [0, 0, 0, 1]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_multiple_controls(self,):
-        circ = TestCircuit(3)
+    def test_cz_multiple_controls(self,):
+        circ = StateVectorCircuit(3)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y,
+            _Z,
             control_qubit=[0, 1], target_qubit=[2], control_value=[1, 1])
 
         expected_gate = np.array([[1, 0, 0, 0, 0, 0, 0, 0],
@@ -113,25 +110,25 @@ class TestTestCircuitCYGate(unittest.TestCase):
                                   [0, 0, 0, 1, 0, 0, 0, 0],
                                   [0, 0, 0, 0, 1, 0, 0, 0],
                                   [0, 0, 0, 0, 0, 1, 0, 0],
-                                  [0, 0, 0, 0, 0, 0, 0, -1j],
-                                  [0, 0, 0, 0, 0, 0, 1j, 0]])
+                                  [0, 0, 0, 0, 0, 0, 1, 0],
+                                  [0, 0, 0, 0, 0, 0, 0, -1]])
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
-    def test_cy_multiple_targets(self,):
+    def test_cz_multiple_targets(self,):
 
-        circ = TestCircuit(3)
+        circ = StateVectorCircuit(3)
         gate_0 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[1, 2], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[1, 2], control_value=[1]
         )
 
-        circ = TestCircuit(3)
+        circ = StateVectorCircuit(3)
         gate_1_0 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[1], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[1], control_value=[1]
         )
         gate_1_1 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _Y, control_qubit=[0], target_qubit=[2], control_value=[1]
+            _Z, control_qubit=[0], target_qubit=[2], control_value=[1]
         )
 
         self.assertIsNone(
