+++ test/simulator/state_vector_circuit/test_single_qubit_gate.py
@@ -2,15 +2,17 @@ import unittest
 
 import numpy as np
 
-from quantestpy import TestCircuit, test_circuit
+from quantestpy import StateVectorCircuit
+from quantestpy.simulator import state_vector_circuit
 
 
-class TestTestCircuitSingleQubitGate(unittest.TestCase):
+class TestStateVectorCircuitSingleQubitGate(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_test_circuit_single_qubit_gate
+    $ python -m unittest \
+        test.simulator.state_vector_circuit.test_single_qubit_gate
 
     ........
     ----------------------------------------------------------------------
@@ -21,8 +23,8 @@ class TestTestCircuitSingleQubitGate(unittest.TestCase):
     """
 
     def test_h_to_0_in_qubit_2(self,):
-        h = test_circuit._H
-        circ = TestCircuit(2)
+        h = state_vector_circuit._H
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
             gate=h, control_qubit=[], target_qubit=[0], control_value=[]
         )
@@ -36,8 +38,8 @@ class TestTestCircuitSingleQubitGate(unittest.TestCase):
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_h_to_1_in_qubit_2(self,):
-        h = test_circuit._H
-        circ = TestCircuit(2)
+        h = state_vector_circuit._H
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
             gate=h, control_qubit=[], target_qubit=[1], control_value=[]
         )
@@ -51,8 +53,8 @@ class TestTestCircuitSingleQubitGate(unittest.TestCase):
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_x_to_0_and_1_in_qubit_2(self,):
-        x = test_circuit._X
-        circ = TestCircuit(2)
+        x = state_vector_circuit._X
+        circ = StateVectorCircuit(2)
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
             gate=x, control_qubit=[], target_qubit=[0, 1], control_value=[]
         )
@@ -66,8 +68,8 @@ class TestTestCircuitSingleQubitGate(unittest.TestCase):
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_h_to_0_and_1_in_qubit_3_qiskit_convention(self,):
-        h = test_circuit._H
-        circ = TestCircuit(3)
+        h = state_vector_circuit._H
+        circ = StateVectorCircuit(3)
         circ._from_right_to_left_for_qubit_ids = True
         actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
             gate=h, control_qubit=[], target_qubit=[0, 1], control_value=[]
